qencode/requirements.txt
pennylane
black
pytorch